[feat] Implement a policy according to the number of accesses
This policy takes into account the total number of accesses for a cached
RDD. We use teraflag to locate the object's total number of accesses,
object group id, and the object's state. When Spark access an RDD, we
reduce the number of accesses for this object. Then we move to TeraCache
only objects whose remaining number of accesses is lower than a
predefined threshold.